feat: changing the way the hidden messages deals with the settings
